requirements.txt
torch>=2.0.0
transformers>=4.37.0
qwen-vl-utils>=0.0.1
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
tiktoken>=0.5.0

